+++ DbMigrate/Model/Support/MigrationRepoMakeDbVersionAware.cs
@@ -22,8 +22,8 @@ namespace DbMigrate.Model.Support
 		{
 			const string createVersionInfoTableSql =
 				@"create table __database_info(
-  min_version_number int,
-  max_version_number int
+  min_version_number long,
+  max_version_number long
 );
 insert into __database_info(min_version_number, max_version_number) values(0, 0);";
 			const string dropVersionInfoTableSql = "drop table __database_info;";
